chore: complete full refactor
