+++ package.json
@@ -16,14 +16,15 @@
     "gcp-build": "npm run build",
     "generate:env": "terraform -chdir=infra apply -target=local_file.env_file",
     "logs": "./bin/get-logs.sh",
+    "logs:watch": "./bin/watch-logs.sh",
     "prestart": "npm run build",
     "start": "NODE_ENV=development functions-framework --target=governanceWatchdog",
     "test": "npm run test:local:ProposalCreated && npm run test:local:ProposalQueued && npm run test:local:ProposalExecuted && npm run test:local:ProposalCanceled && npm run test:local:healthcheck",
-    "test:local:ProposalCanceled": "curl -H \"Content-Type: application/json\" -d @src/proposal-canceled/fixture.json localhost:8080",
-    "test:local:ProposalCreated": "curl -H \"Content-Type: application/json\" -d @src/proposal-created/fixture.json localhost:8080",
-    "test:local:ProposalExecuted": "curl -H \"Content-Type: application/json\" -d @src/proposal-executed/fixture.json localhost:8080",
-    "test:local:ProposalQueued": "curl -H \"Content-Type: application/json\" -d @src/proposal-queued/fixture.json localhost:8080",
-    "test:local:healthcheck": "curl -H \"Content-Type: application/json\" -d @src/health-check/fixture.json localhost:8080",
+    "test:local:ProposalCanceled": "curl -H \"Content-Type: application/json\" -d @src/events/fixtures/proposal-canceled.fixture.json localhost:8080",
+    "test:local:ProposalCreated": "curl -H \"Content-Type: application/json\" -d @src/events/fixtures/proposal-created.fixture.json localhost:8080",
+    "test:local:ProposalExecuted": "curl -H \"Content-Type: application/json\" -d @src/events/fixtures/proposal-executed.fixture.json localhost:8080",
+    "test:local:ProposalQueued": "curl -H \"Content-Type: application/json\" -d @src/events/fixtures/proposal-queued.fixture.json localhost:8080",
+    "test:local:healthcheck": "curl -H \"Content-Type: application/json\" -d @src/events/fixtures/health-check.fixture.json localhost:8080",
     "test:prod:ProposalCanceled": "./bin/test-deployed-function.sh ProposalCanceled",
     "test:prod:ProposalCreated": "./bin/test-deployed-function.sh ProposalCreated",
     "test:prod:ProposalExecuted": "./bin/test-deployed-function.sh ProposalExecuted",
