+++ src/event-notifications/send-discord-notification.ts
@@ -1,6 +1,6 @@
 import { EmbedBuilder, WebhookClient } from "discord.js";
-import getNotificationChannels from "./utils/get-notification-channels.js";
-import getSecret from "./utils/get-secret.js";
+import config from "../config.js";
+import getSecret from "../utils/get-secret.js";
 
 /**
  * Generic Discord notification function that can be reused by different event handlers
@@ -12,7 +12,18 @@ export default async function sendDiscordNotification(
   content: string,
   embed: EmbedBuilder,
 ) {
-  const { discordWebhookUrlSecretId } = getNotificationChannels();
+  // Inline notification channel selection logic
+  const isDevelopment = process.env.NODE_ENV === "development";
+
+  if (isDevelopment) {
+    if (!config.DISCORD_TEST_WEBHOOK_URL_SECRET_ID) {
+      throw new Error("DISCORD_TEST_WEBHOOK_URL_SECRET_ID env var is not set");
+    }
+  }
+
+  const discordWebhookUrlSecretId = isDevelopment
+    ? config.DISCORD_TEST_WEBHOOK_URL_SECRET_ID
+    : config.DISCORD_WEBHOOK_URL_SECRET_ID;
 
   await new WebhookClient({
     url: await getSecret(discordWebhookUrlSecretId),
