+++ src/event-notifications/send-telegram-notification.ts
@@ -1,23 +1,29 @@
-import config from "./config.js";
-import getNotificationChannels from "./utils/get-notification-channels.js";
-import getSecret from "./utils/get-secret.js";
-import { createFormattedMessage } from "./utils/telegram.js";
+import config from "../config.js";
+import getSecret from "../utils/get-secret.js";
 
 /**
- * Sends a notification to Telegram with the provided message data
- * @param title The title for the notification message
- * @param msgData Key-value data to be included in the notification
+ * Sends a pre-formatted HTML message to Telegram
+ * @param formattedMessage HTML-formatted message string ready for Telegram
  * @returns Promise that resolves when the notification is sent
  */
 export default async function sendTelegramNotification(
-  title: string,
-  msgData: Record<string, string | number>,
+  formattedMessage: string,
 ) {
   const botToken = await getSecret(config.TELEGRAM_BOT_TOKEN_SECRET_ID);
   const botUrl = `https://api.telegram.org/bot${botToken}/sendMessage`;
 
-  const formattedMessage = createFormattedMessage(title, msgData);
-  const { telegramChatId } = getNotificationChannels();
+  // Inline notification channel selection logic
+  const isDevelopment = process.env.NODE_ENV === "development";
+
+  if (isDevelopment) {
+    if (!config.TELEGRAM_TEST_CHAT_ID) {
+      throw new Error("TELEGRAM_TEST_CHAT_ID env var is not set");
+    }
+  }
+
+  const telegramChatId = isDevelopment
+    ? config.TELEGRAM_TEST_CHAT_ID
+    : config.TELEGRAM_CHAT_ID;
 
   const payload = {
     chat_id: telegramChatId,
