+++ src/events/fixtures/health-check.fixture.json
