+++ src/events/fixtures/proposal-canceled.fixture.json
