+++ src/events/fixtures/proposal-created.fixture.json
