--- src/proposal-canceled/index.ts
@@ -1,4 +1,0 @@
-import { createEventHandler } from "../utils/event-handler.js";
-import { proposalCanceledConfig } from "./config.js";
-
-export default createEventHandler(proposalCanceledConfig);
