--- src/proposal-created/index.ts
@@ -1,4 +1,0 @@
-import { createEventHandler } from "../utils/event-handler.js";
-import { proposalCreatedConfig } from "./config.js";
-
-export default createEventHandler(proposalCreatedConfig);
