--- src/proposal-executed/index.ts
@@ -1,4 +1,0 @@
-import { createEventHandler } from "../utils/event-handler.js";
-import { proposalExecutedConfig } from "./config.js";
-
-export default createEventHandler(proposalExecutedConfig);
