--- src/proposal-queued/index.ts
@@ -1,4 +1,0 @@
-import { createEventHandler } from "../utils/event-handler.js";
-import { proposalQueuedConfig } from "./config.js";
-
-export default createEventHandler(proposalQueuedConfig);
