+++ src/utils/parse-request-body.ts
@@ -1,6 +1,5 @@
-import { EventType, QuicknodeEvent } from "./types.js";
-import getProposalTimelockId from "./utils/get-proposal-time-lock-id.js";
-import isValidQuicknodePayload from "./utils/is-valid-quicknode-payload.js";
+import { EventType, QuicknodeEvent } from "../events/types.js";
+import isValidQuicknodePayload from "./is-valid-quicknode-payload.js";
 
 /**
  * Parse request body containing parsed events from QuickNode
@@ -8,7 +7,6 @@ import isValidQuicknodePayload from "./utils/is-valid-quicknode-payload.js";
 export default function parseRequestBody(
   requestBody: unknown,
 ): QuicknodeEvent[] {
-  // Validate that the request body has the expected structure
   if (!isValidQuicknodePayload(requestBody)) {
     throw new Error(
       `Request body is not a valid QuickNode payload: ${JSON.stringify(
@@ -20,26 +18,16 @@ export default function parseRequestBody(
   const parsedEvents: QuicknodeEvent[] = [];
 
   for (const event of requestBody.result) {
-    // Check if the event name is a valid EventType
     const eventType = Object.values(EventType).includes(event.name as EventType)
       ? (event.name as EventType)
       : EventType.Unknown;
 
     if (eventType === EventType.Unknown) {
-      if (process.env.DEBUG) {
-        console.log(`Skipping unknown event: '${event.name}'.`);
-      }
-      // Skip events we're not interested in
+      console.log(`Skipping unknown event: '${event.name}'.`);
       continue;
     }
 
-    parsedEvents.push({
-      ...event,
-      timelockId:
-        event.name === EventType.ProposalCreated
-          ? getProposalTimelockId(event)
-          : undefined,
-    });
+    parsedEvents.push(event);
   }
 
   return parsedEvents;
